requirements.txt
dj-database-url==0.4.2
Django==2.0.1
django-crispy-forms==1.7.0
gunicorn==19.7.1
Pillow==5.0.0
psycopg2==2.7.3.1
python-decouple==3.1
pytz==2017.2
whitenoise==3.3.1
